INFR-103: Tighten parameters/arguments agreement
This commit changes the default Raymond rules regarding the necessary
agreement between parameters and arguments, and treatment of the
optional `*Options` parameter. In the new rule, a final `*Options`
parameter must be `*Options` and cannot be captured through
`interface{}`. The number of remaining parameters and arguments must
agree. Rules for agreement of a single parameter/argument pair were
not modified.

Also, this commit makes this repository a proper go module and brings
in the mustache submodule files into the current repository so that
tests can be executed.